Move EventQueue to event module and add event-queue-api feature flag
