+++ Cargo.toml
@@ -65,6 +65,10 @@ docs-only = ["neon-runtime/docs-only"]
 # Feature flag to enable the try_catch API of RFC 29.
 try-catch-api = []
 
+# Feature flag to enable the `EventQueue` API of RFC 33.
+# https://github.com/neon-bindings/rfcs/pull/32
+event-queue-api = []
+
 # Feature flag to include procedural macros
 proc-macros = ["neon-macros"]
