+++ src/context/mod.rs
@@ -14,8 +14,8 @@ use borrow::{Ref, RefMut, Borrow, BorrowMut};
 use borrow::internal::Ledger;
 use context::internal::Env;
 use handle::{Managed, Handle};
-#[cfg(feature = "napi-4")]
-use task::EventQueue;
+#[cfg(all(feature = "napi-4", feature = "event-queue-api"))]
+use event::EventQueue;
 use types::{JsValue, Value, JsObject, JsArray, JsFunction, JsBoolean, JsNumber, JsString, StringResult, JsNull, JsUndefined};
 #[cfg(feature = "napi-1")]
 use types::boxed::{Finalize, JsBox};
@@ -382,7 +382,7 @@ pub trait Context<'a>: ContextInternal<'a> {
         JsBox::new(self, v)
     }
 
-    #[cfg(feature = "napi-4")]
+    #[cfg(all(feature = "napi-4", feature = "event-queue-api"))]
     /// Creates an unbounded queue of events to be executed on a JavaScript thread
     fn queue(&mut self) -> EventQueue {
         EventQueue::new(self)
@@ -605,7 +605,7 @@ impl<'a> TaskContext<'a> {
         })
     }
 
-    #[cfg(feature = "napi-4")]
+    #[cfg(all(feature = "napi-4", feature = "event-queue-api"))]
     pub(crate) fn with_context<T, F: for<'b> FnOnce(TaskContext<'b>) -> T>(env: Env, f: F) -> T {
         Scope::with(env, |scope| {
             f(TaskContext { scope })
