+++ src/task/mod.rs
@@ -1,11 +1,5 @@
 //! Utilities for scheduling tasks to be executed by the Node.js runtime
 
-#[cfg(feature = "napi-4")]
-mod event_queue;
-
-#[cfg(feature = "napi-4")]
-pub use self::event_queue::{EventQueue, EventQueueError};
-
 use std::marker::{Send, Sized};
 use std::mem;
 use std::os::raw::c_void;
